--- config/snake_config_model_fao.yml
@@ -1,75 +1,0 @@
-
-###########################################################################################
-#################################### General settings ######################################
-
-# Path to where the wflow model will be stored (include the project name)
-project_dir: tests/test_project
-
-# Path to the data catalog to use (can be different for Linux/Windows/data stored on P drive or other)
-data_sources: 
-  - artifact_data
-  - tests/data/tests_data_catalog.yml
-
-# Path to the data catalog to use (can be different for Linux/Windows/data stored on P drive or other)
-data_sources_climate: config/cmip6_data.yml
-
-# Info on the historical climate data is used in the historic model run + reference data for the weather generator
-# Time horizon of the historical period
-starttime: "2000-01-01T00:00:00"
-endtime: "2010-12-31T00:00:00"
-
-# Historical climate data source (name as available in the data_sources catalog file). Either [era5, chirps_global, chirps, eobs].
-#default to set up the model
-clim_historical: 
-  - era5
-  - chirps_global
-
-climate_sources_colors:
-  - blue
-  - green 
-
-# List of wflow output variables to save
-wflow_outvars: ['river discharge', 'precipitation', 'overland flow', 'actual evapotranspiration', 'groundwater recharge', 'snow']
-
-###########################################################################################
-#################################### Model building #######################################
-
-# hydromt "region" for building the model, coordinates are in EPSG 4326
-# for more info on how to fill this, you can see the hydromt documentation: 
-# https://deltares.github.io/hydromt/latest/user_guide/cli.html#region-options
-# For wflow please use either 'basin' or 'subbasin' keywords
-model_region: "{'subbasin': [12.3758, 46.4076], 'uparea': 70}"
-# Grid resolution of the model in degrees
-model_resolution: 0.0166666
-# Path to model build configuration
-model_build_config: "config/fao/wflow_build_model.yml"
-waterbodies_config: "config/fao/wflow_update_waterbodies.yml"
-
-# If available, please here put the path to the available observation stations locations files
-# csv file containing three columns: station_ID, x, y
-output_locations: "tests/data/observations/discharge-locations-grdc.csv" # If set to None, will default to the wflow outlets
-# Path to the file with observed discharge timeseries per station 
-observations_timeseries: "tests/data/observations/observations-discharge-grdc.csv"
-
-###########################################################################################
-################################## Climate projections ####################################
-# climate models and scenarios
-clim_project: cmip6
-#models: ['NOAA-GFDL/GFDL-ESM4', 'INM/INM-CM4-8' ,'INM/INM-CM5-0','NIMS-KMA/KACE-1-0-G','NCC/NorESM2-MM','NCC/NorESM2-LM','CMCC/CMCC-CM2-SR5','CMCC/CMCC-ESM2']
-models: ['NOAA-GFDL/GFDL-ESM4', 'CMCC/CMCC-ESM2', 'INM/INM-CM5-0']
-#scenarios: [ssp126, ssp245, ssp370, ssp585]
-scenarios: [ssp245, ssp585]
-members: [r1i1p1f1]
-
-variables: [precip, temp]
-
-# time horizons of the climate projections
-start_month_hyd_year: Jan
-historical: 1980, 2010
-future_horizons: 
-  near: 2030, 2060
-  far: 2070, 2100
-
-# Additional gridded outputs (else only basin average)
-save_grids: FALSE
-
