+++ src/setup_time_horizon.py
@@ -11,7 +11,6 @@ def prep_hydromt_update_forcing_config(
     endtime: str,
     fn_yml: Union[str, Path] = "wflow_build_forcing_historical.yml",
     precip_source: str = "era5",
-    suffix: bool = True,
 ):
     """Prepare a hydromt config file to be able to add forcing to a wflow model
 
@@ -25,8 +24,6 @@ def prep_hydromt_update_forcing_config(
         Path to the output hydromt config file
     precip_source : str
         Name of the precipitation source to use
-    suffix: bool
-        add {precip_source} as suffix to config file name, path_forcing name and run_default name, default is True
     """
     # Check precip source and set options accordingly
     if precip_source == "eobs":
@@ -39,16 +36,11 @@ def prep_hydromt_update_forcing_config(
         pet_method = "debruin"
     # TODO: make more flexible to allow using temp variables if they would be available in the dataset.
 
-    if suffix == True:
-        path_forcing = (
-            f"../climate_historical/wflow_data/inmaps_historical_{precip_source}.nc"
-        )
-        config_name = f"wflow_sbm_{precip_source}.toml"
-        dir_output = f"run_default_{precip_source}"
-    else:
-        path_forcing = "../climate_historical/wflow_data/inmaps_historical.nc"
-        config_name = "wflow_sbm.toml"
-        dir_output = "run_default"
+    path_forcing = (
+        f"../climate_historical/wflow_data/inmaps_historical_{precip_source}.nc"
+    )
+    config_name = f"wflow_sbm_{precip_source}.toml"
+    dir_output = f"run_default_{precip_source}"
 
     forcing_options = {
         "setup_config": {
@@ -90,7 +82,6 @@ if __name__ == "__main__":
             endtime=sm.params.endtime,
             fn_yml=sm.output.forcing_yml,
             precip_source=sm.params.clim_source,
-            suffix=sm.params.suffix,
         )
     else:
         prep_hydromt_update_forcing_config(
@@ -98,5 +89,4 @@ if __name__ == "__main__":
             endtime="2010-12-31T00:00:00",
             fn_yml="wflow_build_forcing_historical.yml",
             precip_source="era5",
-            suffix=False,
         )
